PP_3.1.5 rest

+++ src/main/java/ru/kata/spring/boot_security/demo/services/RoleServiceImpl.java
@@ -32,12 +32,12 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public Role findById(long id) {
+    public Role findById(int id) {
         return roleDao.findById(id);
     }
 
     @Override
-    public List<Role> findByIdRoles(List<Long> rolesId) {
+    public List<Role> findByIdRoles(List<Integer> rolesId) {
         return roleDao.findByIdRoles(rolesId);
     }
